Added file custom_filters
Added file with custom filters. Updated the examples
